working PoC with hardcoded file names and paths
